ai_toolbox_project/public/Games/QuizzerAI/python-quiz-app-main/requirements.txt
flask
gunicorn
altgraph==0.17.4
astroid==3.3.8
cfgv==3.4.0
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
filelock==3.16.1
identify==2.6.3
isort==5.13.2
macholib==1.16.3
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.0.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pylint==3.3.3
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==75.6.0
tomlkit==0.13.2
virtualenv==20.28.0
